requirements.txt
Django==2.2
django-cors-headers==2.5.2
django-extensions==2.1.6
djangorestframework==3.9.2
celery==4.3.0

